requirements_v2.txt
# yomiageBotEx v2 最小限の依存関係

# Discord
py-cord[voice]>=2.6.0

# 設定・環境変数
PyYAML>=6.0
python-dotenv>=1.0.0

# HTTP通信
aiohttp>=3.9.0

# 音声処理（録音用）
PyNaCl>=1.5.0